Added working text to speech and Eye tracker mouse
If you are intrested the tests 1-7 are what I used as the version control for the eye tracker, as it took a while and I didn't want to lose the progress. Cole did the speech to text quickly so it was uploaded at once. The complete set of code it title part1.py and part2.py where part one is the working eye tracker to mouse code and part two is the working speech to text.